fix VC compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@5118 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/Enviro/wx/TerrManDlg.cpp
@@ -64,7 +64,7 @@ void TerrainManagerDlg::RefreshTreeContents()
 	m_pTree->DeleteAllItems();
 
 	vtStringArray &paths = m_DataPaths;
-	unsigned int i, num = paths.size();
+	size_t i, num = paths.size();
 	wxString2 wstr, wstr2;
 
 	m_Root = m_pTree->AddRoot(_("Terrain Data Paths"));
